Do not access specific pointers when non existent
Make sure we do not access coordinates of specific
pointers before mOngoingPinchResize is true.

Removed the unnecessary log messages as well, since
the InputEvent object is logged already and ACTION_MOVE
MotionEvent will be logged once mOngoingPinchResize is true
in onPinchResize.

This should fix the IllegalArgumentException that was emerging earlier

Bug: 271894660
Test: Reproduce the steps in the bug

Change-Id: I6e570c5381ebf98a7b7b271acdabaa99aadf9e5d