Improve PowerGroup event notification
When the default display group is asleep/dozing and some other display
in another group becomes awake, we were sending incorrect and
misleading information from PowerManagerService to
PhoneWindowManager. In particular, we were conflating the "global
wakefulness" with the wakefulness of the default display, which led to
problems like the default display waking up when it shouldn't and
Always On Display ceasing to work.

This CL refactors things to make the distinction more clear between
global wakefulness changes and specific display PowerGroup changes.

Bug: 255688811
Test: atest PowerManagerServiceTest
Change-Id: I809d5fcd8eaa61dc2d07b63a0ae3b5aa924faa1c