Fix the hidden task was removed before starting home
The home activity idle event could be calls after switching
to other tasks. In that case, it can unexpectedly remove the
hidden tasks that were added after leaving home.

Now, we remove the hidden tasks if home is idle and resumed
to ensure the hidden tasks are safe to be removed.

Also fixing an issue that the task affinity of the
singleInstance activity should be unique.

Bug: 264602067
Test: atest RecentTasksTest
Change-Id: Idb4c71ebbc54b3b37197e0b0c1c049dd85fd6f44