[multi-shade] Fixes issue where NSWV was still visible after unlock.
For reasons that are not yet understood, the NotificationShadeWindowView
remaines visible after unlocking the device.

The solution here simply removes the bouncer-showing logic from our
MultiShadeMotionEventInteractor but keeps the logic that checks the
direction of drag. If the user drags up the screen, our system does not
intercept the touch, allowing the legacy code to show the bouncer in its
own, correct way - which gets rid of the NSWV visibility bug. If the
user drags down the screen, we intercept touch and show the shade(s).

Bug: 274159734
Flag: DUAL_SHADE
Test: unit tests updated
Test: manually verified that, after unlocking the device, touch goes
through and I can click app icons on the launcher. Also made sure that
dragging up on the lock screen brings up the bouncer and I can unlock
the device from it.

Change-Id: I9f0d02f403007946b6911a5057fac9ffa316a9c7