Layer "locked" methods for better locking logs.
When reporting either deadlocks or lock contention, the OS logs the
line at which the lock was acquired, and our LocalHandler makes this
tricky to track across builds as line numbers change.

This change pivots to adding a call stack frame to give ourselves
clearer information about what operation is working with the lock.

Also restore missing code coverage for health checks that was
accidentally deleted.

Bug: 274681945
Test: atest FrameworksMockingServicesTests:BroadcastQueueTest
Test: atest FrameworksMockingServicesTests:BroadcastQueueModernImplTest
Test: atest FrameworksMockingServicesTests:BroadcastRecordTest
Change-Id: Ic6de76f9a7e1ae7b9d1bd6358f556072eaf3ac7c