Send HOME Intent for the main display of the visibule user, with the visible userId.
Previously CATEGORY_HOME is used for the default display, and
CATEGORY_SECONDARY_HOME is used for all non-default displays with
the current user id. But, in the concurrent multi user enviorment,
we should use CATEGORY_HOME Intent for each user's main display,
CATEGORY_SECONDARY_HOME for the others.

This CL fixes
- to use CATEGORY_HOME for main displays assigned to concurrent users.
- to send HOME Intent with the correct userId.

Bug: 272281432
Test: atest MultiDisplaySystemDecorationTests
Change-Id: I562e271b7b350f06238f6291ed84c468cfb2d28a